+++ xephas_app/CHANGELOG.md
@@ -1,6 +1,11 @@
-# 0.3.0+3
+# 0.4.0+4
 
 - Latest version of `xephas_app` brick
+- This now includes custom lints for riverpod and the app in `analysis_options.yaml`
+
+## 0.3.0+3
+
+- Minor changes to the initial version of `xephas_app` brick
 
 ## 0.1.0+1